requirements.txt
flask
openai-whisper
pytubefix
bert-extractive-summarizer
torch
transformers
